Migrate to the SPIRE Plugin SDK
This change migrates the plugin to use the SPIRE Plugin SDK added for
the SPIRE 1.0.0 release. Legacy plugins (those that don't use the SDK)
are no longer supported in SPIRE 1.1.0.

The changes are mostly mechanical as far as the consuming the plugin
SDK. Changes to tests were a little more involved, particularly removing
the dependencies on the SPIRE codebase that pulled in a bunch of code
from the agent unnecessarily.

The `tpm: ` error prefix was removed from all errors produced by the
plugin as SPIRE now automatically adds the plugin name as a prefix on
all errors produced from plugins.

The small test helper was added to exercise the node attestation flow.
It is generically useful and a good fit for the SDK. I'll open a PR
there in the near future to add it and at that point open a second PR
here to consume it from the SDK and remove the local copy.

Signed-off-by: Andrew Harding <aharding@vmware.com>